[279. Perfect Squares][Compile Error]committed by LeetCode Extension

+++ 279-Perfect-Squares/solution.java
@@ -2,7 +2,7 @@ public class Solution {
     public int numSquares(int n) {
         int[] result = new int[n + 1];
         for(int i=1; i<=n; i++){
-            int tmp = Math.sqrt(n);
+            int tmp = (int) Math.sqrt(n);
             if(n == tmp * tmp){
                 result[i] = 1;
             }else{
